+++ qcb-firmware/inc/qcfp.h
@@ -24,9 +24,10 @@ SOFTWARE.
 
 #include "qcb.h"
 
+// Corresponds to the max size of the buffers in the event queue
 #define MAX_QCFP_PACKET_SIZE 32
 
 void qcfp_init(void);
-void qcfp_data_received(uint8_t buffer[], uint16_t buffer_size);
+void qcfp_data_received(uint8_t buffer[], uint8_t buffer_size);
 
 #endif // _QCFP_H_
